+++ examples/realworld.svelte.dev/src/routes/auth/user.js
@@ -1,0 +1,5 @@
+export function get(req, res) {
+	res.setHeader('Content-Type', 'application/json');
+
+	res.end(JSON.stringify({ user: req.session.user || null }));
+}