+++ examples/realworld.svelte.dev/src/routes/index.svelte
@@ -1,0 +1,5 @@
+<script>
+	import Home from '$components/Home.svelte';
+</script>
+
+<Home p={1}/>