+++ examples/realworld.svelte.dev/src/routes/login/index.svelte
@@ -1,0 +1,62 @@
+<script context="module">
+	export async function preload({ params }, { user }) {
+		if (user) {
+			this.redirect(302, `/`);
+		}
+	}
+</script>
+
+<script>
+	import { goto, stores } from '$app/client.js';
+	import ListErrors from '$components/ListErrors.svelte';
+	import { post } from '$common/utils.js';
+
+	const { session } = stores();
+
+	let email = '';
+	let password = '';
+	let errors = null;
+
+	async function submit(event) {
+		const response = await post(`auth/login`, { email, password });
+
+		// TODO handle network errors
+		errors = response.errors;
+
+		if (response.user) {
+			$session.user = response.user;
+			goto('/');
+		}
+	}
+</script>
+
+<svelte:head>
+	<title>Sign in • Conduit</title>
+</svelte:head>
+
+<div class="auth-page">
+	<div class="container page">
+		<div class="row">
+			<div class="col-md-6 offset-md-3 col-xs-12">
+				<h1 class="text-xs-center">Sign In</h1>
+				<p class="text-xs-center">
+					<a href="/register">Need an account?</a>
+				</p>
+
+				<ListErrors {errors}/>
+
+				<form on:submit|preventDefault={submit}>
+					<fieldset class="form-group">
+						<input class="form-control form-control-lg" type="email" required placeholder="Email" bind:value={email}>
+					</fieldset>
+					<fieldset class="form-group">
+						<input class="form-control form-control-lg" type="password" required placeholder="Password" bind:value={password}>
+					</fieldset>
+					<button class="btn btn-lg btn-primary pull-xs-right" type="submit">
+						Sign in
+					</button>
+				</form>
+			</div>
+		</div>
+	</div>
+</div>