+++ examples/sandbox/src/routes/index.svelte
@@ -1,0 +1,60 @@
+<script context="module">
+	export function preload(page, session) {
+		return {
+			answer: session.answer
+		};
+	}
+</script>
+
+<script>
+	import Counter from '$components/Counter.svelte';
+
+	export let answer;
+</script>
+
+<main>
+	<h1>Hello world!</h1>
+
+	<p>the answer is {answer}</p>
+
+	<Counter/>
+	<p>Visit the <a href="https://svelte.dev">svelte.dev</a> to learn how to build Svelte apps.</p>
+</main>
+
+<style>
+	:root {
+		font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen, Ubuntu, Cantarell, 'Open Sans', 'Helvetica Neue', sans-serif;
+	}
+
+	main {
+		text-align: center;
+		padding: 1em;
+		margin: 0 auto;
+	}
+
+	h1 {
+		color: #ff3e00;
+		text-transform: uppercase;
+		font-size: 4rem;
+		font-weight: 100;
+		line-height: 1.1;
+		margin: 4rem auto;
+		max-width: 14rem;
+	}
+
+	p {
+		max-width: 14rem;
+		margin: 2rem auto;
+		line-height: 1.35;
+	}
+
+	@media (min-width: 480px) {
+		h1 {
+			max-width: none;
+		}
+
+		p {
+			max-width: none;
+		}
+	}
+</style>