+++ packages/app-utils/src/render/index.ts
@@ -1,7 +1,7 @@
 import { createHash } from 'crypto';
 import render_page from './page';
 import render_route from './route';
-import { IncomingRequest, RenderOptions } from '../types';
+import { EndpointResponse, IncomingRequest, PageResponse, RenderOptions } from '../types';
 
 function md5(body) {
 	return createHash('md5').update(body).digest('hex');
@@ -10,24 +10,39 @@ function md5(body) {
 export async function render(
 	request: IncomingRequest,
 	options: RenderOptions
-) {
-	const response = await (
-		render_route(request, options) ||
-		render_page(request, options)
-	);
+): Promise<EndpointResponse | PageResponse> {
+	const { context, headers = {} } = (await options.setup.prepare?.(request.headers)) || {};
 
-	// inject ETags for 200 responses
-	if (response && response.status === 200) {
-		if (!/(no-store|immutable)/.test(response.headers['cache-control'])) {
-			const etag = `"${md5(response.body)}"`;
+	try {
+		const response = await (
+			render_route(request, context, options) ||
+			render_page(request, context, options)
+		);
 
-			if (request.headers['if-none-match'] === etag) {
-				return { status: 304 };
+		if (response) {
+			// inject ETags for 200 responses
+			if (response.status === 200) {
+				if (!/(no-store|immutable)/.test(response.headers['cache-control'])) {
+					const etag = `"${md5(response.body)}"`;
+
+					if (request.headers['if-none-match'] === etag) {
+						return { status: 304 };
+					}
+
+					response.headers['etag'] = etag;
+				}
 			}
 
-			response.headers['etag'] = etag;
+			return {
+				status: response.status,
+				headers: { ...headers, ...response.headers },
+				body: response.body
+			};
 		}
+	} catch (err) {
+		return {
+			status: 500,
+			body: options.dev ? err.stack : err.message
+		};
 	}
-
-	return response;
 }