+++ packages/app-utils/src/render/page.ts
@@ -18,6 +18,7 @@ type FetchOpts = {
 
 export default async function render_page(
 	request: IncomingRequest,
+	context: any,
 	options: RenderOptions
 ) {
 	let redirected;
@@ -27,7 +28,7 @@ export default async function render_page(
 
 	const baseUrl = ''; // TODO
 
-	const session = {}; // TODO
+	const session = await options.setup.getSession?.(context);
 
 	const serialized_session = try_serialize(session, err => {
 		throw new Error(`Failed to serialize session data: ${err.message}`);