+++ packages/adapter-netlify/CHANGELOG.md
@@ -1,6 +1,20 @@
 # @sveltejs/adapter-netlify
 
+## 0.0.10
+
+### Patch Changes
+
+- Use setup
+
+## 0.0.9
+
+### Patch Changes
+
+- 0320208: Rename 'server route' to 'endpoint'
+- 5ca907c: Use shared mkdirp helper
+
 ## 0.0.8
+
 ### Patch Changes
 
 - various