+++ packages/adapter-node/CHANGELOG.md
@@ -1,5 +1,34 @@
 # @sveltejs/adapter-node
 
+## 0.0.11
+
+### Patch Changes
+
+- Use setup module
+- Updated dependencies [undefined]
+- Updated dependencies [undefined]
+- Updated dependencies [undefined]
+  - @sveltejs/app-utils@0.0.14
+
+## 0.0.10
+
+### Patch Changes
+
+- Updated dependencies [undefined]
+- Updated dependencies [96c06d8]
+  - @sveltejs/app-utils@0.0.13
+
+## 0.0.9
+
+### Patch Changes
+
+- 0320208: Rename 'server route' to 'endpoint'
+- 026acd2: Accommodate missing build/prerendered directory
+- 8b63057: Let sirv add must-revalidate
+- Updated dependencies [0320208]
+- Updated dependencies [5ca907c]
+  - @sveltejs/app-utils@0.0.12
+
 ## 0.0.8
 
 ### Patch Changes